requirements.txt
streamlit==1.28.0
neo4j==5.14.0
elasticsearch==8.11.0
openai==1.3.0
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0
pyvis==0.3.2
streamlit-autorefresh==0.0.1
python-dotenv==1.0.0
